Updated model.
