+++ systems/data-plan/package/pkg/config.go
@@ -13,7 +13,8 @@ type Config struct {
 	Queue            *uconf.Queue     `default:"{}"`
 	Timeout          time.Duration    `default:"3s"`
 	MsgClient        *uconf.MsgClient `default:"{}"`
-	Service          *uconf.Service
+	Service          *uconf.Service   `default:"{}"`
+	Rate             *uconf.Service   `default:"{}"`
 }
 
 func NewConfig(name string) *Config {
@@ -25,6 +26,9 @@ func NewConfig(name string) *Config {
 		MsgClient: &uconf.MsgClient{
 			Timeout:        5 * time.Second,
 			ListenerRoutes: []string{"event.cloud.dataplan.baserate.upload"},
+		},
+		Rate: &uconf.Service{
+			Uri: "rate:9090",
 		},
 	}
 }
